+++ src/core/interfaces/client.ts
@@ -1,0 +1,5 @@
+import type { SendRequestOptions } from './sendRequestOptions';
+
+export interface Client {
+  sendRequest<T>(options: SendRequestOptions<T>): Promise<T>;
+}