+++ src/core/interfaces/index.ts
@@ -1,0 +1,5 @@
+export type { Client } from './client';
+
+export type { ClientConfig } from './clientConfig';
+
+export type { SendRequestOptions } from './sendRequestOptions';